Fixed difference between ending time
If the forge data ends before the time period on the map file, then the last forge time period will be extend to end on the day the map says the scenario end. 

There might be problems if the map requires the forge file to be shortened by more than the duration on the final forge period.