Updated dependencies, dropped excel dupe, fixes..
Migrated to spork 0.2.1.0-SNAPSHOT, also
updated to clojure 1.9 (which caught some errors),
added proc 0.2.6 (wasn't in dependencies for some
reason), and updated incanter to 1.9.3 (for clj 1.9
compatibility and other fixes).

Migrated existing demand_builder.m4plugin code
to use new API in spork.util.excel.core.

Fixed error in demand_builder.formatter that
was a faulty (and undetected in clj 1.8)
application of destructuring to a map.