MAP extension fixed for uneeded SRCs
Some SRCs become uneeded for the scenario in the forge file after a
certain point.  We don't want to extend these to the end of the
period specified by the MAP file if the SRC doesn't have a demand
active on the last day.  To do this, we need to check for a demand on the
last day-not just a demand in the last phase.
For the demands that end short, their duration had to be dec'ed.