updated test cases to match new format
Changed the forge files to follow the Unit_Node_Detail format instead of the SRC_By_Day format. 
In the actual data there will be more unused columns. 
Updated the headers on the consolidated and map file to match exactly. 
When reading data directly from the excel file, will automatically format to have the expected headers (these test on test using the txt files)

Some of the expected results need to be updated. There are some off-by-one issues because the forge data starts on day 1 but the expect data assumes it starts on day 0.