Updated doc, schema, and typo
For FORGE Unit_Node_Detail sheets, the Branch Code field is a mix
of some strings and some numbers.  If the first record contains a
number and no schema was supplied, the spork parser will assume that
the entire field consists of numbers.  This will throw an exception
if it runs into non-numeric character.  Fix was to supply a schema
when converting the sheet to records.  The Branch Code field isn't
even used in processing, so we don't need it.
Also noticed that the global start and end times for each demand
group in the demand group shifter were not right due to a typo in
the sorting function.