Vignettes being excluded erroneously
We were filtering out the records for all SRCs but the first one
from the vignette consolidated file.  We want to keep the vignette
demand for all of the SRCs specified in the consolidated file.

Since Unit_Node_Detail i/o is currently not working per the open
issue, we default to using the SRC_By_Day sheet for now.